Added '`offset`' capability to cursor/keyboard rotation as an alternative to '`setNorth`' Fixed bug where flinging during cursor/keyboard rotation would start pan fling animation and exit rotation Refactored `InteractionOptions` into seperate source file
