+++ lib/src/map/options/options.dart
@@ -9,6 +9,7 @@ import 'package:flutter_map/src/gestures/multi_finger_gesture.dart';
 import 'package:flutter_map/src/map/camera/camera_constraint.dart';
 import 'package:flutter_map/src/map/camera/camera_fit.dart';
 import 'package:flutter_map/src/map/inherited_model.dart';
+import 'package:flutter_map/src/map/options/interaction.dart';
 import 'package:flutter_map/src/misc/fit_bounds_options.dart';
 import 'package:flutter_map/src/misc/position.dart';
 import 'package:flutter_map/src/misc/private/positioned_tap_detector_2.dart';
@@ -34,8 +35,6 @@ typedef PointerHoverCallback = void Function(
   PointerHoverEvent event,
   LatLng point,
 );
-
-typedef IsKeyboardKeyTrigger = bool Function(LogicalKeyboardKey key)?;
 
 @immutable
 class MapOptions {
@@ -345,151 +344,4 @@ class MapOptions {
         maxBounds,
         interactionOptions,
       ]);
-}
-
-@immutable
-final class InteractionOptions {
-  /// See [InteractiveFlag] for custom settings
-  final int flags;
-
-  /// Prints multi finger gesture winner Helps to fine adjust
-  /// [rotationThreshold] and [pinchZoomThreshold] and [pinchMoveThreshold]
-  /// Note: only takes effect if [enableMultiFingerGestureRace] is true
-  final bool debugMultiFingerGestureWinner;
-
-  /// If true then [rotationThreshold] and [pinchZoomThreshold] and
-  /// [pinchMoveThreshold] will race If multiple gestures win at the same time
-  /// then precedence: [pinchZoomWinGestures] > [rotationWinGestures] >
-  /// [pinchMoveWinGestures]
-  final bool enableMultiFingerGestureRace;
-
-  /// Rotation threshold in degree default is 20.0 Map starts to rotate when
-  /// [rotationThreshold] has been achieved or another multi finger gesture wins
-  /// which allows [MultiFingerGesture.rotate] Note: if [interactiveFlags]
-  /// doesn't contain [InteractiveFlag.rotate] or [enableMultiFingerGestureRace]
-  /// is false then rotate cannot win
-  final double rotationThreshold;
-
-  /// When [rotationThreshold] wins over [pinchZoomThreshold] and
-  /// [pinchMoveThreshold] then [rotationWinGestures] gestures will be used. By
-  /// default only [MultiFingerGesture.rotate] gesture will take effect see
-  /// [MultiFingerGesture] for custom settings
-  final int rotationWinGestures;
-
-  /// Pinch Zoom threshold default is 0.5 Map starts to zoom when
-  /// [pinchZoomThreshold] has been achieved or another multi finger gesture
-  /// wins which allows [MultiFingerGesture.pinchZoom] Note: if
-  /// [interactiveFlags] doesn't contain [InteractiveFlag.pinchZoom] or
-  /// [enableMultiFingerGestureRace] is false then zoom cannot win
-  final double pinchZoomThreshold;
-
-  /// When [pinchZoomThreshold] wins over [rotationThreshold] and
-  /// [pinchMoveThreshold] then [pinchZoomWinGestures] gestures will be used. By
-  /// default [MultiFingerGesture.pinchZoom] and [MultiFingerGesture.pinchMove]
-  /// gestures will take effect see [MultiFingerGesture] for custom settings
-  final int pinchZoomWinGestures;
-
-  /// Pinch Move threshold default is 40.0 (note: this doesn't take any effect
-  /// on drag) Map starts to move when [pinchMoveThreshold] has been achieved or
-  /// another multi finger gesture wins which allows
-  /// [MultiFingerGesture.pinchMove] Note: if [interactiveFlags] doesn't contain
-  /// [InteractiveFlag.pinchMove] or [enableMultiFingerGestureRace] is false
-  /// then pinch move cannot win
-  final double pinchMoveThreshold;
-
-  /// When [pinchMoveThreshold] wins over [rotationThreshold] and
-  /// [pinchZoomThreshold] then [pinchMoveWinGestures] gestures will be used. By
-  /// default [MultiFingerGesture.pinchMove] and [MultiFingerGesture.pinchZoom]
-  /// gestures will take effect see [MultiFingerGesture] for custom settings
-  final int pinchMoveWinGestures;
-
-  @Deprecated(
-    'Prefer `flags.scrollWheelZoom`. '
-    'This property was moved as it better suited being an `InteractiveFlag`. '
-    'This property is deprecated since v6.',
-  )
-  final bool enableScrollWheel;
-
-  final double scrollWheelVelocity;
-
-  /// Whether to allow rotation by moving the cursor dependent on the currently
-  /// pressed keyboard [LogicalKeyboardKey]
-  ///
-  /// Fix to returning `false` to disable cursor/keyboard rotation.
-  ///
-  /// Defaults to allowing rotation by cursor if any of the Control keys are
-  /// pressed.
-  final IsKeyboardKeyTrigger isCursorRotationKeyboardKeyTrigger;
-
-  const InteractionOptions({
-    this.flags = InteractiveFlag.all,
-    this.debugMultiFingerGestureWinner = false,
-    this.enableMultiFingerGestureRace = false,
-    this.rotationThreshold = 20.0,
-    this.rotationWinGestures = MultiFingerGesture.rotate,
-    this.pinchZoomThreshold = 0.5,
-    this.pinchZoomWinGestures =
-        MultiFingerGesture.pinchZoom | MultiFingerGesture.pinchMove,
-    this.pinchMoveThreshold = 40.0,
-    this.pinchMoveWinGestures =
-        MultiFingerGesture.pinchZoom | MultiFingerGesture.pinchMove,
-    this.enableScrollWheel = true,
-    this.scrollWheelVelocity = 0.005,
-    this.isCursorRotationKeyboardKeyTrigger,
-  })  : assert(
-          rotationThreshold >= 0.0,
-          'rotationThreshold needs to be a positive value',
-        ),
-        assert(
-          pinchZoomThreshold >= 0.0,
-          'pinchZoomThreshold needs to be a positive value',
-        ),
-        assert(
-          pinchMoveThreshold >= 0.0,
-          'pinchMoveThreshold needs to be a positive value',
-        );
-
-  bool get dragEnabled => InteractiveFlag.hasDrag(flags);
-
-  bool get flingEnabled => InteractiveFlag.hasFlingAnimation(flags);
-
-  bool get doubleTapZoomEnabled => InteractiveFlag.hasDoubleTapZoom(flags);
-
-  bool get rotateEnabled => InteractiveFlag.hasRotate(flags);
-
-  bool get pinchZoomEnabled => InteractiveFlag.hasPinchZoom(flags);
-
-  bool get pinchMoveEnabled => InteractiveFlag.hasPinchMove(flags);
-
-  @override
-  bool operator ==(Object other) =>
-      other is InteractionOptions &&
-      flags == other.flags &&
-      debugMultiFingerGestureWinner == other.debugMultiFingerGestureWinner &&
-      enableMultiFingerGestureRace == other.enableMultiFingerGestureRace &&
-      rotationThreshold == other.rotationThreshold &&
-      rotationWinGestures == other.rotationWinGestures &&
-      pinchZoomThreshold == other.pinchZoomThreshold &&
-      pinchZoomWinGestures == other.pinchZoomWinGestures &&
-      pinchMoveThreshold == other.pinchMoveThreshold &&
-      pinchMoveWinGestures == other.pinchMoveWinGestures &&
-      // ignore: deprecated_member_use_from_same_package
-      enableScrollWheel == other.enableScrollWheel &&
-      scrollWheelVelocity == other.scrollWheelVelocity;
-
-  @override
-  int get hashCode => Object.hash(
-        flags,
-        debugMultiFingerGestureWinner,
-        enableMultiFingerGestureRace,
-        rotationThreshold,
-        rotationWinGestures,
-        pinchZoomThreshold,
-        pinchZoomWinGestures,
-        pinchMoveThreshold,
-        pinchMoveWinGestures,
-        // ignore: deprecated_member_use_from_same_package
-        enableScrollWheel,
-        scrollWheelVelocity,
-      );
 }
